TEST: change recv logic
When client close the socket, it send FIN/ACK packet to server.
When server receive this packet, this data lenght is zero.
So, when recv size is 0 means that client close the socket.
But if server limit the maximum of the client numbers, it handle with accept queue.
When client leave in accept queue before accept() function call,
Recv will fail because can't read anything from the client socket.
So, server MUST handle when recv size is below than 0